use config setup for all tests
