+++ packages/css/src/test/types.test.ts
@@ -1,34 +1,5 @@
-import { css, createConfig, createCss } from '../';
-
-/* -------------------------------------------------------------------------------------------------
- * test setup
- * -----------------------------------------------------------------------------------------------*/
-
-const config = createConfig({
-  include: ['./src/**/*.{ts,tsx}'],
-  responsive: {
-    md: '@media (min-width: 700px)',
-    lg: '@media (min-width: 1024px)',
-    xl: '@media (min-width: 1280px)',
-    '2xl': '@media (min-width: 1536px)',
-  },
-  theme: {
-    color: {
-      'slate-100': '#f1f5f9',
-      'slate-700': '#334155',
-      'sky-500': '#0ea5e9',
-    },
-  },
-  aliases: {
-    p: ['padding'],
-  },
-});
-
-type Config = typeof config;
-
-declare module '@tokenami/dev' {
-  interface TokenamiConfig extends Config {}
-}
+import { css, createCss } from '../';
+import { config } from './setup';
 
 /* -------------------------------------------------------------------------------------------------
  * test cases
